added edit option
can edit your own posts

+++ FrontEnd/newquery.php
@@ -5,10 +5,17 @@
     $user_id = $_SESSION['user_id'];
 
     if(isset($_GET['modifypost'])){
-
         $post_id = $_GET['post_id'];
         $post = new Post;
-        $post->deleteUserPost($post_id,$user_id);
+        if($_GET['modifypost'] === 'Edit'){
+           
+            
+        }
+
+        if($_GET['modifypost'] === 'Delete'){
+            $post->deleteUserPost($post_id,$user_id);
+        }
+      
     }
 
     if(isset($_GET['submitpost'])){
@@ -16,10 +23,18 @@
         $title = $_GET['title'];
         $imglink = $_GET['img-link'];
         $txtarea = $_GET['txt-area'];
-     
         $data = array($title,$imglink,$txtarea,$user_id);
         $post = new Post;
-        $post->addUserPost($data);
+
+        if($_GET['submitpost'] === 'Submit'){
+            $post->addUserPost($data);
+        }
+
+        if($_GET['submitpost'] === 'Edit'){
+            $post->updateUserPost($data);
+        }
+
+       
     }
 
     header('Location: ' . $_SERVER['HTTP_REFERER']);
